Store wheel in docker container
We use a workaround to store the wheel corresponding to the master branch.
This is because we cannot use the job token to download artifacts
using the gitlab API with gitlab CE (premium feature).
Hence the trick to store the wheel in a docker image (:-().
this is transitory before the package is released on PyPi.